requirements.txt
python-telegram-bot==20.3
langchain
faiss-cpu
transformers
sentence-transformers
torch
